+++ hutool-ai/src/main/java/cn/hutool/ai/core/AIConfigRegistry.java
@@ -19,7 +19,6 @@ package cn.hutool.ai.core;
 import cn.hutool.core.util.ServiceLoaderUtil;
 
 import java.util.Map;
-import java.util.ServiceLoader;
 import java.util.concurrent.ConcurrentHashMap;
 
 /**
@@ -34,9 +33,12 @@ public class AIConfigRegistry {
 
 	// 加载所有 AIConfig 实现类
 	static {
-		final ServiceLoader<AIConfig> loader = ServiceLoaderUtil.load(AIConfig.class);
-		for (final AIConfig config : loader) {
-			configClasses.put(config.getModelName().toLowerCase(), config.getClass());
+		for (final AIConfig config : ServiceLoaderUtil.load(AIConfig.class)) {
+			configClasses.putIfAbsent(config.getModelName().toLowerCase(), config.getClass());
+		}
+		// issue#4241@github，多线程和Spring环境下可能导致SPI文件找不到问题
+		for (final AIConfig config : ServiceLoaderUtil.load(AIConfig.class, AIConfig.class.getClassLoader())) {
+			configClasses.putIfAbsent(config.getModelName().toLowerCase(), config.getClass());
 		}
 	}