+++ hutool-core/src/main/java/cn/hutool/core/convert/impl/BeanConverter.java
@@ -79,7 +79,7 @@ public class BeanConverter<T> extends AbstractConverter<T> {
 
 		if(value instanceof Map ||
 				value instanceof ValueProvider ||
-				BeanUtil.isBean(value.getClass())) {
+				BeanUtil.isReadableBean(value.getClass())) {
 			if(value instanceof Map && this.beanClass.isInterface()) {
 				// 将Map动态代理为Bean
 				return MapProxy.create((Map<?, ?>)value).toProxyBean(this.beanClass);